Remove no-npm option from release workflow
Not useful as part of the workflow, the most important part of the
workflow is publishing to npmjs.com, only here we can add provenance.

Because the option will also skip the `npm version` and build commands,
it won't help in case we only need to retry failed git/github related
release instructions. But if everything related to git or github fails
after publishing the npm package all of these can easily be retried from
a local machine.