Overhaul linting + formatting approach
The standard linter didn't play nice with Prettier, so that we had
standard undo certain formatting applied by Prettier. Thus we couldn't
rely on the Prettier extension in say VS Code to apply formatting right
away upon saving: an extra call in the CLI was necessary to apply the
final formatting..

This change simplifies the approach: we're using Prettier for formatting
with an externalized config so that editors can pick it up, and eslint
for linting, and the two won't interfere.

This also allows us to add explicit steps for linting and formatting
checks in the CI pipeline.